+++ src/interceptor/twcc/estimator/delay_based/overuse_detector.rs
@@ -1,33 +1,33 @@
 use super::*;
 
 struct ArrivalTimeFilter {
-    m_hat: f32,
-    e: f32,
-    var_v_hat: f32,
+    m_hat: f64,
+    e: f64,
+    var_v_hat: f64,
 }
 
 impl ArrivalTimeFilter {
     fn new(intergroup_delay: i64) -> ArrivalTimeFilter {
         ArrivalTimeFilter {
-            m_hat: intergroup_delay as f32,
+            m_hat: intergroup_delay as f64,
             e: INITIAL_SYSTEM_ERROR_COVARIANCE,
             var_v_hat: 0.0,
         }
     }
 
     /// Updates the Kalman filter, returning the previous intergroup delay variation estimate.
-    fn update(&mut self, intergroup_delay: i64, min_send_interval: i64) -> f32 {
+    fn update(&mut self, intergroup_delay: i64, min_send_interval: i64) -> f64 {
         // This is different than in the Google CC alg. draft since the interval used here is in
         // microseconds
-        let alpha = (1.0 - CHI).powf(30.0 * (min_send_interval as f32) / 1e6);
+        let alpha = (1.0 - CHI).powf(30.0 * (min_send_interval as f64) / 1e6);
 
         let q = STATE_NOISE_COVARIANCE;
 
         // m_hat, intergroup_delay, and z are in microseconds
-        let z = intergroup_delay as f32 - self.m_hat;
+        let z = intergroup_delay as f64 - self.m_hat;
         let z2 = z * z;
 
-        self.var_v_hat = f32::max(1.0, alpha * self.var_v_hat + (1.0 - alpha) * z2);
+        self.var_v_hat = f64::max(1.0, alpha * self.var_v_hat + (1.0 - alpha) * z2);
         let k = (self.e + q) / (self.var_v_hat + (self.e + q));
         self.e = (1.0 - k) * (self.e + q);
 
@@ -38,7 +38,7 @@ impl ArrivalTimeFilter {
 }
 
 struct DelayThreshold {
-    threshold: f32,
+    threshold: f64,
 }
 
 impl DelayThreshold {
@@ -48,8 +48,8 @@ impl DelayThreshold {
         }
     }
 
-    fn update(&mut self, interarrival_time: i64, intergroup_delay_estimate: f32) {
-        let interarrival_time = interarrival_time as f32;
+    fn update(&mut self, interarrival_time: i64, intergroup_delay_estimate: f64) {
+        let interarrival_time = interarrival_time as f64;
 
         let threshold_delta = intergroup_delay_estimate.abs() - self.threshold;
         if threshold_delta <= 15000.0 {
