+++ src/interceptor/twcc/estimator/loss_based.rs
@@ -7,9 +7,9 @@ impl LossBasedBandwidthEstimator {
         LossBasedBandwidthEstimator {}
     }
 
-    pub fn estimate(&mut self, current_bandwidth: f32, received: u32, lost: u32) -> f32 {
+    pub fn estimate(&mut self, current_bandwidth: f64, received: u32, lost: u32) -> f64 {
         let total = received + lost;
-        let fraction_lost = lost as f32 / total as f32;
+        let fraction_lost = lost as f64 / total as f64;
         if fraction_lost < 0.02 {
             current_bandwidth * 1.05
         } else if fraction_lost > 0.10 {
